Add quick-sort algorithm
The fastest sorting algorithm, one of the top 10 algorithms of 20th
century. It just sorts the data in place in linearithmic expected time.
Although the worst case of quick-sort running time is quadratic,
this implementation employed randomized partitioning that reduces a
chance of such a worst case (its still may happen but its more likely to
be eaten by shark in a desert). Other one drawback of this algorithm is
that it is not stable (read it as it confuses items when they were sorted by
some other comparing strategy)

**Important notes**
 - Implementation doesn't use recursion, because I despise it

Implementation was borrowed from [Cormen's book](http://mitpress.mit.edu/books/introduction-algorithms-third-edition)