Add insertion-sort algorithm
Insertion sort is simple one and although its upper bound running time
is quadratic, if its is sorting almost sorted array it produces linear time!
This sorting algorithm is a first to chose when you are dealing with not
very long or partially sorted sequences, because in this case it
outperformes other algorithms. And yes, it is stable!

Implementation was borrowed from [Cormen's
book](http://mitpress.mit.edu/books/introduction-algorithms-third-edition)