Add Graph data-structures
Graph is a collection of vertices and edges between them. This is very
abstract data-structure that can  represent plenty of concepts. For
example: road map is a graph where roads are edges and joints are
vertices, social relationship is a graph where people are vertices and
their relations are edges