Implement Topological sort algorithm
Topological sort is somehow different from an original sense of sorting.
It is sorting vertices in a DAG (Directed Acyclic Graph) in a way of
precedence (e.g suppose we have a lot of tasks. Some of them  might be
done separately,but some of them should be done only after others is
done. Applying this sort we can get a consecutive sequence of tasks
which we may perform without being blocked)

Again, implementation was borrowed from [Cormen's  book](http://mitpress.mit.edu/books/introduction-algorithms-third-edition)