Implement AVL Tree
AVL Tree is a data-structure that represents self balancing
binary-search tree that was broadly used before the invention of
Red Black Tree. Its idea is to keep additional information about height
in each node and do not allow sibling nodes to be different in heights
on more then 1. This property is done using rotations on
insertion and deletion operations. An AVL Tree is guaranteed to have
height at most logN and provide basic dictionary operations in a
logarithmic complexity