Improve performance of max square sums algorithm
Algorithm consist of 4 parts: check on answer 1, on 2, 3, or 4.
Checking on 2 is the most time consuming part. It takes o(NlogN) time
and O(N) space (finds prime factors of a number). Moving this part to
the end drastically improves an algorithm in practice. However upper
bound of NlogN still holds